chore: update structures
