+++ shared/components/timetable/eventCard/contactsCard.tsx
@@ -1,10 +1,10 @@
 /// <reference types="vite-plugin-svgr/client" />
 import React from 'react';
-import AvatarIcon from '../../assets/avatar.svg?react';
+import AvatarIcon from '../../../assets/avatar.svg?react';
 
 const ContactsCard = ({ organizer }: { organizer?: string }) => {
   return (
-    <div>
+    <>
       <h2 className="mt-7">Контакты:</h2>
       <h3 className="mt-1">Организаторы/ спикеры</h3>
       <div className="container flex flex-row flex-nowrap items-center pb-1.5">
@@ -14,7 +14,7 @@ const ContactsCard = ({ organizer }: { organizer?: string }) => {
           <p className="c1">{organizer ?? 'Неизвестно'}</p>
         </div>
       </div>
-    </div>
+    </>
   );
 };
