+++ shared/components/timetable/eventCard/detailsCard.tsx
@@ -1,15 +1,10 @@
 import React from 'react';
-import Button from '../common/button';
+import Button from '../../common/button.tsx';
 
 const DetailsCard = ({ link }: { link?: string }) => {
   if (!link) return <div className="mt-3" />;
   return (
-    <Button
-      className="mt-5"
-      onClick={() => {
-        window.open(link, '_blank');
-      }}
-    >
+    <Button className="mt-5" onClick={() => window.open(link, '_blank')}>
       <h5 className="underline">Подробности мероприятия</h5>
     </Button>
   );
