+++ shared/components/timetable/eventCard/viewMapCard.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Button from '../common/button';
+import Button from '../../common/button.tsx';
 
 const ViewMapCard = ({ link }: { link?: string }) => {
   return (
