+++ web/src/pages/timetable/eventDescription.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Layout from '~/../../shared/components/common/layout';
 
-import EventCard from '~/../../shared/components/timetable/eventCard.tsx';
+import EventCard from '~/components/timetable/eventCard';
 import useLoaderData from '~/utils/routerLoader.ts';
 import HeaderBar from '~/widgets/headerBar';
 import NavigationBar from '~/widgets/navigationBar';
